% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddbar.R
\name{ddbar}
\alias{ddbar}
\title{Multilevel Drill Down Bar Widget}
\usage{
ddbar(
  data,
  options = NULL,
  horizontal = F,
  showTitle = T,
  reactiveID = NULL,
  delimiter = "|",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{data list object to be plotted; typically generated using \code{dataFormat()}.}

\item{options}{list of apache echarts options}

\item{horizontal}{bool indicating whether the bar plot should be horizontal}

\item{showTitle}{bool indicating whether to display a dynamic title}

\item{reactiveID}{string id to be passed to \code{Shiny.setInputValue()} called within ddbar.js}

\item{delimiter}{string delimiter for the dataGroupId; must match the delimiter used by \code{dataFormat()}.}

\item{width}{width of the graph as CSS units.}

\item{height}{height of the graph as CSS units.}

\item{elementId}{html element id.}
}
\description{
Multilevel Drill Down Bar Widget
}
\examples{

rawdata <- data.frame(nationality = sample(c("French", "German", "British"),
                      100, replace=TRUE, prob=c(0.4, 0.3, 0.3)),
                      sex = sample(c("Male", "Female"),
                      100, replace=TRUE, prob=c(0.5, 0.5)),
                      occupation = sample(c("Chef", "Pilot", "Developer"),
                      100, replace=TRUE, prob=c(0.1, 0.7, 0.2)),
                      netWorth = rpois(100, 1.5) * 100000)

options <- list(
  tooltip = list(
    trigger = 'axis',
    axisPointer = list(
      type= 'shadow'
    )
  )
)

rawdata |>
  dataFormat(mean, mode = "vector") |>
  ddbar(options)

}
