% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFormat.R
\name{handleMissingData}
\alias{handleMissingData}
\title{Handles missing values in the input dataframe}
\usage{
handleMissingData(data, na.action)
}
\arguments{
\item{data}{input dataframe}

\item{na.action}{string indicating how missing data should be handled. "omit" or "label" options may be used.}
}
\description{
Handles missing values in the input dataframe
}
