% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddbar.R
\name{ddbar}
\alias{ddbar}
\title{Drill Down Bar Apache echarts Widget}
\usage{
ddbar(data, width = NULL, height = NULL, elementId = NULL)
}
\description{
\if{html}{\out{
<Add Description>
}}
}
\examples{

data = list(list(dataGroupId = "1",
                 data = list(c("1_1", 5, "1_1"),
                             c("1_2", 2, "1_2"))),
            list(dataGroupId = "1_1",
                 data = list(c("1_1_1", 6, "1_1_1"),
                             c("1_1_2", 8, "1_1_2"))),
            list(dataGroupId = "1_2",
                 data = list(c("1_2_1", 1),
                             c("1_2_2", 3))),
            list(dataGroupId = "1_1_1",
                 data = list(c("1_1_1_1", 10),
                             c("1_1_1_2", 9))),
            list(dataGroupId = "1_1_2",
                 data = list(c("1_1_2_1", 12),
                             c("1_1_2_2", 7)))
       )

ddbar(data)

}
