% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddbar.R
\name{ddbar}
\alias{ddbar}
\title{Drill Down Bar Apache echarts Widget}
\usage{
ddbar(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{data list object to be plotted; typically generated using \code{dataFormat()}.}

\item{width}{width of the graph as CSS units.}

\item{height}{height of the graph as CSS units.}

\item{elementId}{html element id.}
}
\description{
Drill Down Bar Apache echarts Widget
}
\examples{

rawdata <- data.frame(nationality = sample(c("French", "German", "British"), 100, replace=TRUE, prob=c(0.4, 0.3, 0.3)),
                      sex = sample(c("Male", "Female"), 100, replace=TRUE, prob=c(0.5, 0.5)),
                      occupation = sample(c("Chef", "Pilot", "Developer"), 100, replace=TRUE, prob=c(0.1, 0.7, 0.2)),
                      netWorth = rpois(100, 1.5) * 100000)

rawdata |>
  dataFormat(mean) |>
  ddbar()

}
